Stop publishing synthetic Godeps.json files
From https://github.com/kubernetes/enhancements/blob/master/keps/sig-architecture/917-go-modules/README.md#publish-staging-component-modules-to-individual-repositories:
> Generate synthetic Godeps.json files containing the SHA or git tag of
> module dependencies, for consumption by downstream consumers using
> dependency management tools like glide. Continue publishing these at
> least until our minimum supported version of go defaults to enabling
> module support (currently targeted for go 1.13, which is approximately
> Kubernetes 1.16-1.17 timeframe).

Now that Go 1.16 is in module-aware mode by default, it should be safe
to remove support for publishing synthetic `Godeps.json` files.
