+++ Dockerfile
@@ -36,7 +36,6 @@ ADD _output/collapsed-kube-commit-mapper /collapsed-kube-commit-mapper
 ADD _output/sync-tags /sync-tags
 ADD _output/init-repo /init-repo
 
-ADD _output/godeps-gen /godeps-gen
 ADD _output/gomod-zip /gomod-zip
 ADD artifacts/scripts/ /publish_scripts
