+++ Makefile
@@ -40,7 +40,6 @@ build:
 	$(call build_cmd,publishing-bot)
 	$(call build_cmd,sync-tags)
 	$(call build_cmd,init-repo)
-	$(call build_cmd,godeps-gen)
 	$(call build_cmd,gomod-zip)
 .PHONY: build
